resolver: Add support for excluding global catalogs from resolution
The resolver is currently configured to always consider resolving
from global catalogs. In some cases, this may not be
desirable, and the user would like to explicitly ignore global catalogs
for the purposes of resolution. This change enables per-namespace
exclusion from global catalog sources via an annotation on the
operatorgroup.

Signed-off-by: Daniel Sover <dsover@redhat.com>